[win32] powermanagement: move implementation to platform folder
