+++ cmake/treedata/windows/subdirs.txt
@@ -1,17 +1,17 @@
-xbmc/platform/win32             platform/win32
-xbmc/platform/win32/input       platform/win32/input
-xbmc/platform/win32/filesystem  platform/win32/filesystem
-xbmc/platform/win32/network     platform/win32/network
-xbmc/platform/win32/peripherals platform/win32/peripherals
-xbmc/platform/win32/storage     platform/win32/storage
-xbmc/input/touch                input/touch
-xbmc/input/touch/generic        input/touch/generic
-xbmc/network/mdns               network/mdns
-xbmc/powermanagement/windows    powermanagement/windows
-xbmc/utils/win32                utils/win32
-xbmc/rendering/dx               rendering/dx
-xbmc/threads/platform/win       threads/win
-xbmc/windowing/windows          windowing/windows
+xbmc/platform/win32                    platform/win32
+xbmc/platform/win32/input              platform/win32/input
+xbmc/platform/win32/filesystem         platform/win32/filesystem
+xbmc/platform/win32/network            platform/win32/network
+xbmc/platform/win32/peripherals        platform/win32/peripherals
+xbmc/platform/win32/powermanagement    platform/win32/powermanagement
+xbmc/platform/win32/storage            platform/win32/storage
+xbmc/input/touch                       input/touch
+xbmc/input/touch/generic               input/touch/generic
+xbmc/network/mdns                      network/mdns
+xbmc/utils/win32                       utils/win32
+xbmc/rendering/dx                      rendering/dx
+xbmc/threads/platform/win              threads/win
+xbmc/windowing/windows                 windowing/windows
 xbmc/cores/RetroPlayer/process/windows cores/RetroPlayer/process/windows
 xbmc/cores/RetroPlayer/rendering/VideoShaders/windows cores/RetroPlayer/rendering/VideoShaders/windows
 xbmc/cores/VideoPlayer/Process/windows cores/VideoPlayer/Process/windows
