+++ xbmc/platform/win32/powermanagement/CMakeLists.txt
@@ -3,4 +3,4 @@ set(SOURCES Win32PowerSyscall.cpp
 
 set(HEADERS Win32PowerSyscall.h)
 
-core_add_library(powermanagement_windows)
+core_add_library(platform_win32_powermanagement)
