+++ xbmc/platform/win32/powermanagement/Win32PowerSyscall.h
