+++ xbmc/platform/win32/powermanagement/WinIdleTimer.cpp
